+++ src/benchmarks/gc/GC.Infrastructure/GC.Infrastructure/Commands/ReliabilityFrameworkTest/ReliabilityFrameworkTestCreateTestSuiteCommand.cs
@@ -1,6 +1,4 @@
 ﻿using GC.Infrastructure.Core.Configurations;
-using GC.Infrastructure.Core.Configurations.ReliabilityFrameworkTest;
-using GC.Infrastructure.Core.TaskRunner.ReliabilityFrameworkTest;
 using Spectre.Console;
 using Spectre.Console.Cli;
 using System.ComponentModel;
